# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Trips",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/trips": {
            "get": {
                "summary": "Mostrar todos los viajessss",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los viajes."
                    },
                    "default": {
                        "description": "error al listar viajes."
                    }
                }
            }
        },
        "/api/trips/count": {
            "get": {
                "summary": "Mostrar numero de viajes registrados",
                "operationId": "App\\Http\\Controllers\\TripController::count",
                "responses": {
                    "200": {
                        "description": "Mostrar numero de viajes registrados."
                    },
                    "default": {
                        "description": "error al contar viajes."
                    }
                }
            }
        },
        "/api/trips/count/": {
            "post": {
                "summary": "Mostrar numero de viajes por ciudad",
                "operationId": "App\\Http\\Controllers\\TripController::countByCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Ciudad para filtrar conteo",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contar viajes registrados por ciudad enviada."
                    },
                    "default": {
                        "description": "error al contar viajes por ciudad."
                    }
                }
            }
        },
        "/api/trips/status/{trip_id}": {
            "get": {
                "summary": "Estado de viaje",
                "operationId": "App\\Http\\Controllers\\TripController::statusByTrip",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "id de registro de viaje",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra estado de viaje enviado."
                    },
                    "default": {
                        "description": "error al mostrar status del viaje."
                    }
                }
            }
        },
...
        "/api/trips/{trip_id}/": {
            "get": {
                "summary": "Mostrar un viaje",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "id de viaje que se quiere viusalizar",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string",
                            "format": "string",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visualiza data de viaje seleccionado."
                    },
                    "default": {
                        "description": "Viaje no encontrado."
                    }
                }
            }
        },
...
    }
}